requirements.txt
anthropic>=0.25.0
gradio>=4.31.0
python-dotenv>=1.0.0
